requirements.txt
streamlit==1.35.0
transformers==4.41.2
torch==2.3.0
openai-whisper==20231117
ffmpeg-python==0.2.0
gTTS==2.5.1